Added html-report feature
