+++ package.json
@@ -10,7 +10,8 @@
     "build": "cd client_src && npm run-script build && cd ..",
     "start": "node .",
     "test": "npm run e2etest",
-    "e2etest": "./node_modules/.bin/cucumber-js test/web/features/*.feature --exit"
+    "e2etest": "./node_modules/.bin/cucumber-js test/web/features/*.feature -f json:test/report/cucumber_report.json --exit",
+    "report": "node test/report/report.js"
   },
   "dependencies": {
     "ajv": "^6.5.5",
