+++ test/report/report.js
@@ -1,0 +1,11 @@
+var reporter = require('cucumber-html-reporter');
+ 
+var options = {
+        theme: 'bootstrap',
+        jsonFile: 'test/report/cucumber_report.json',
+        output: 'test/report/cucumber_report.html',
+        reportSuiteAsScenarios: true,
+        launchReport: true
+    };
+ 
+return reporter.generate(options);
